+++ Runtime/NodeTypes/CompositionNode.cs
@@ -5,5 +5,12 @@ namespace BTDesigner
     public abstract class CompositionNode : Node
     {
         public List<Node> children = new();
+        
+        public override Node Clone()
+        {
+            var node = Instantiate(this);
+            node.children = children.ConvertAll(c => c.Clone());
+            return node;
+        }
     }
 }