+++ Runtime/RootNode.cs
@@ -2,10 +2,16 @@ using UnityEngine;
 
 namespace BTDesigner
 {
-    [CreateAssetMenu(fileName = "Start Node", menuName = "BTDesign/Start Node", order = 1)]
     public class RootNode : Node
     {
         [SerializeReference] public Node child;
+
+        public override Node Clone()
+        {
+            RootNode node = Instantiate(this);
+            node.child = child.Clone();
+            return node;
+        }
         
         protected override void OnStart() { }